+- working

+++ LabOOP1/Constatnts/Constants.cs
@@ -2,7 +2,7 @@
 {
     public class Constants
     {
-        public const double ImpVal = 1000000000;
+        public const int ImpVal = 1000000000;
         public const int MaxCountOfFoodStock = 2;
         public const int MaxCountOfResourceStock = 5;
         public const int MaxCountOfFoodStockForHouse = 5;
